requirements.txt
distro==1.9.0
JPype1==1.5.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tabula-py==2.9.0
tqdm==4.66.2
tzdata==2024.1
